+++ simulations/vip-448/bsctestnet.ts
@@ -1,11 +1,11 @@
 import { expectEvents } from "src/utils";
 import { forking, testVip } from "src/vip-framework";
 
-import vip444 from "../../vips/vip-444/bsctestnet";
+import vip448 from "../../vips/vip-448/bsctestnet";
 import OMNICHAIN_PROPOSAL_SENDER_ABI from "./abi/OmnichainProposalSender.json";
 
 forking(47754551, async () => {
-  testVip("VIP-444", await vip444(), {
+  testVip("VIP-448", await vip448(), {
     callbackAfterExecution: async txResponse => {
       await expectEvents(
         txResponse,
